requirements.txt
# Merchant Fraud Detection Simulator
# Python 3.8+ required

# Core data processing
pandas>=1.5.0
numpy>=1.24.0

# Machine learning
scikit-learn>=1.3.0

# Dashboard
streamlit>=1.28.0
plotly>=5.17.0

# Configuration
pyyaml>=6.0.1

# Utilities
python-dateutil>=2.8.2
joblib>=1.3.0
